Split toolopts from cli both in module as code
In the past, we used to have this portion of code all together inside
the toolopts module, making it very difficult to understand and improve
as the separation was not intuitive enough.

With this patch, we are splitting the cli related functions and classes
to its own module, making it an entrypoint for everything related to CLI
usage.

Toolopts also got its own module to further separate the logic between
cli stuff and options stuff. Historically we call all options from our
source code as "toolopts", so to preserve the naming convention and
minimize as much as possible the changes throughout the codebase, the
same name was preserved.

A new addition was also made which is the config module. This module is
responsible to gather all the configuration options we could ever define
for the tool. Currently, it only has CLIConfig and FileConfig, which are
the two supported options.

Toolopts, at this point, is our final class who will hold all the
options that are composed from the cli and config file (being able to be
extended in the future to other options).
